Equalize IAbpLazyServiceProvider and ITransientCachedServiceProvider
To reduce cached service provider service count.
