+++ framework/src/Volo.Abp.Core/Volo/Abp/DependencyInjection/CachedServiceProviderBase.cs
@@ -1,10 +1,9 @@
 using System;
 using System.Collections.Concurrent;
-using Microsoft.Extensions.DependencyInjection;
 
 namespace Volo.Abp.DependencyInjection;
 
-public abstract class CachedServiceProviderBase : IServiceProvider
+public abstract class CachedServiceProviderBase : ICachedServiceProviderBase
 {
     protected IServiceProvider ServiceProvider { get; }
     protected ConcurrentDictionary<Type, Lazy<object>> CachedServices { get; }
@@ -21,6 +20,29 @@ public abstract class CachedServiceProviderBase : IServiceProvider
         return CachedServices.GetOrAdd(
             serviceType,
             _ => new Lazy<object>(() => ServiceProvider.GetService(serviceType))
+        ).Value;
+    }
+    
+    public T GetService<T>(T defaultValue)
+    {
+        return (T)GetService(typeof(T), defaultValue);
+    }
+
+    public object GetService(Type serviceType, object defaultValue)
+    {
+        return GetService(serviceType) ?? defaultValue;
+    }
+
+    public T GetService<T>(Func<IServiceProvider, object> factory)
+    {
+        return (T)GetService(typeof(T), factory);
+    }
+    
+    public object GetService(Type serviceType, Func<IServiceProvider, object> factory)
+    {
+        return CachedServices.GetOrAdd(
+            serviceType,
+            _ => new Lazy<object>(() => factory(ServiceProvider))
         ).Value;
     }
 }
