+++ framework/src/Volo.Abp.Core/Volo/Abp/DependencyInjection/ICachedServiceProvider.cs
@@ -1,6 +1,4 @@
-﻿using System;
-
-namespace Volo.Abp.DependencyInjection;
+﻿namespace Volo.Abp.DependencyInjection;
 
 /// <summary>
 /// Provides services by caching the resolved services.
@@ -8,7 +6,7 @@ namespace Volo.Abp.DependencyInjection;
 /// This service's lifetime is scoped and it should be used
 /// for a limited scope.
 /// </summary>
-public interface ICachedServiceProvider : IServiceProvider
+public interface ICachedServiceProvider : ICachedServiceProviderBase
 {
 
 }
